[refact] navigate 사용을 위한 구조 수정
